Rework Markdown parsing for `BuildCard`
Improves parsing of the commit message inside `BuildCard` including replacing `marked` with a basic custom Markdown parser. Additionally, fixes any issues with spacing in the message leading to incorrect determination of the `title` and `description` and HTML elements such as `<` and `>` which often broke prior leading to broken `Co-Authored-By` statements.